popover in right sidebar too
